+++ packages/metascraper-publisher/index.js
@@ -1,8 +1,7 @@
 'use strict'
 
-const { $jsonld, publisher, toRule } = require('@metascraper/helpers')
+const { $filter, $jsonld, publisher, toRule } = require('@metascraper/helpers')
 
-const REGEX_RSS = /^(.*?)\s[-|]\satom$/i
 const REGEX_TITLE = /^.*?[-|]\s+(.*)$/
 
 const toPublisher = toRule(publisher)
@@ -20,9 +19,8 @@ module.exports = () => ({
     toPublisher($jsonld('publisher.name')),
     toPublisher($ => $('meta[property="og:site_name"]').attr('content')),
     toPublisher($ => $('meta[name*="application-name" i]').attr('content')),
-    toPublisher($ => $('meta[property="al:android:app_name"]').attr('content')),
-    toPublisher($ => $('meta[property="al:iphone:app_name"]').attr('content')),
-    toPublisher($ => $('meta[property="al:ipad:app_name"]').attr('content')),
+    toPublisher($ => $('meta[name*="app-title" i]').attr('content')),
+    toPublisher($ => $('meta[property*="app_name" i]').attr('content')),
     toPublisher($ => $('meta[name="publisher" i]').attr('content')),
     toPublisher($ => $('meta[name="twitter:app:name:iphone"]').attr('content')),
     toPublisher($ =>
@@ -38,15 +36,14 @@ module.exports = () => ({
     toPublisher($ =>
       $('meta[property="twitter:app:name:googleplay"]').attr('content')
     ),
-    toPublisher($ => $('#logo').text()),
-    toPublisher($ => $('.logo').text()),
-    toPublisher($ => $('a[class*="brand" i]').text()),
-    toPublisher($ => $('[class*="brand" i]').text()),
+    toPublisher($ => $filter($, $('#logo'))),
+    toPublisher($ => $filter($, $('.logo'))),
+    toPublisher($ => $filter($, $('a[class*="brand" i]'))),
+    toPublisher($ => $filter($, $('[class*="brand" i]'))),
     toPublisher($ => $('[class*="logo" i] a img[alt]').attr('alt')),
     toPublisher($ => $('[class*="logo" i] img[alt]').attr('alt')),
-    toPublisher($ => getFromTitle($('title').text(), REGEX_TITLE)),
     toPublisher($ =>
-      getFromTitle($('link[type*="xml" i]').attr('title'), REGEX_RSS)
+      $filter($, $('title'), el => getFromTitle($filter.fn(el), REGEX_TITLE))
     )
   ]
 })