+++ packages/metascraper-title/index.js
@@ -11,9 +11,9 @@ module.exports = () => ({
     toTitle($ => $('meta[property="twitter:title"]').attr('content')),
     toTitle($ => $filter($, $('title'))),
     toTitle($jsonld('headline')),
-    toTitle($ => $('.post-title').text()),
+    toTitle($ => $filter($, $('.post-title'))),
     toTitle($ => $filter($, $('.entry-title'))),
-    toTitle($ => $('h1[class*="title" i] a').text()),
-    toTitle($ => $('h1[class*="title" i]').text())
+    toTitle($ => $filter($, $('h1[class*="title" i] a'))),
+    toTitle($ => $filter($, $('h1[class*="title" i]')))
   ]
 })