+++ packages/metascraper/test/integration/qz/index.js
@@ -2,6 +2,7 @@
 
 const snapshot = require('snap-shot')
 const { resolve } = require('path')
+const { omit } = require('lodash')
 const { readFile } = require('fs').promises
 
 const metascraper = require('../../..')([
@@ -16,13 +17,14 @@ const metascraper = require('../../..')([
   require('metascraper-publisher')(),
   require('metascraper-title')(),
   require('metascraper-url')(),
-  require('metascraper-readability')()
+  require('metascraper-readability')(),
+  require('metascraper-audio')()
 ])
 
-const url = 'http://gent.ilcore.com/2012/06/better-timer-for-javascript.html'
+const url = 'https://qz.com/on/beyond-silicon-valley/'
 
-it('fastersite', async () => {
+it('qz', async () => {
   const html = await readFile(resolve(__dirname, 'input.html'))
   const metadata = await metascraper({ html, url })
-  snapshot(metadata)
+  snapshot(omit(metadata, ['date']))
 })