+++ packages/metascraper-image/CHANGELOG.md
@@ -3,6 +3,142 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.22.11](https://github.com/microlinkhq/metascraper/compare/v5.22.10...v5.22.11) (2021-07-19)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.9](https://github.com/microlinkhq/metascraper/compare/v5.22.8...v5.22.9) (2021-07-11)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.8](https://github.com/microlinkhq/metascraper/compare/v5.22.7...v5.22.8) (2021-07-10)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.7](https://github.com/microlinkhq/metascraper/compare/v5.22.6...v5.22.7) (2021-07-08)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.5](https://github.com/microlinkhq/metascraper/compare/v5.22.4...v5.22.5) (2021-06-29)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.4](https://github.com/microlinkhq/metascraper/compare/v5.22.3...v5.22.4) (2021-06-27)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.2](https://github.com/microlinkhq/metascraper/compare/v5.22.1...v5.22.2) (2021-06-23)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.22.1](https://github.com/microlinkhq/metascraper/compare/v5.22.0...v5.22.1) (2021-06-09)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+# [5.22.0](https://github.com/microlinkhq/metascraper/compare/v5.21.10...v5.22.0) (2021-05-31)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.10](https://github.com/microlinkhq/metascraper/compare/v5.21.9...v5.21.10) (2021-05-24)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.9](https://github.com/microlinkhq/metascraper/compare/v5.21.8...v5.21.9) (2021-05-21)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.7](https://github.com/microlinkhq/metascraper/compare/v5.21.6...v5.21.7) (2021-04-24)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.6](https://github.com/microlinkhq/metascraper/compare/v5.21.5...v5.21.6) (2021-04-23)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.5](https://github.com/microlinkhq/metascraper/compare/v5.21.4...v5.21.5) (2021-04-05)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.4](https://github.com/microlinkhq/metascraper/compare/v5.21.3...v5.21.4) (2021-03-30)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.3](https://github.com/microlinkhq/metascraper/compare/v5.21.2...v5.21.3) (2021-03-21)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
+## [5.21.2](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-image/compare/v5.21.1...v5.21.2) (2021-03-15)
+
+**Note:** Version bump only for package metascraper-image
+
+
+
+
+
 # [5.21.0](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-image/compare/v5.20.0...v5.21.0) (2021-03-10)